e2e: use podman-remote in story_openshift scenarios
Involves moving podman command code (test side) from test/integration
to test/extended. Now it can be shared between e2e and integration
tests. This only affects `test/integration/podman_test.go` on the
integration side.

Former feature file tags @story_registry, @story_health, and
@story_marketplace now tag respective scenarios inside
@story_openshift scope.
