+++ test/e2e/testsuite/testsuite.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/crc-org/crc/pkg/crc/constants"
 	"github.com/crc-org/crc/pkg/crc/preset"
+	"github.com/crc-org/crc/test/extended/crc/cmd"
 	crcCmd "github.com/crc-org/crc/test/extended/crc/cmd"
 	"github.com/crc-org/crc/test/extended/util"
 	"github.com/cucumber/godog"
@@ -324,12 +325,18 @@ func InitializeScenario(s *godog.ScenarioContext) {
 		ExecuteCommandWithExpectedExitStatus)
 	s.Step(`^execut(?:e|ing) single crc (.*) command (.*)$`,
 		ExecuteSingleCommandWithExpectedExitStatus)
+	s.Step(`^execut(?:e|ing) podman command (.*) (succeeds|fails)$`,
+		ExecutingPodmanCommandSucceedsFails)
 	s.Step(`^ensuring CRC cluster is running (succeeds|fails)$`,
 		EnsureCRCIsRunningSucceedsOrFails)
-	s.Step(`ensuring user is logged in (succeeds|fails)`,
+	s.Step(`^ensuring user is logged in (succeeds|fails)`,
 		EnsureUserIsLoggedIntoClusterSucceedsOrFails)
+	s.Step(`^podman command is available$`,
+		PodmanCommandIsAvailable)
 	s.Step(`^deleting a pod (succeeds|fails)$`,
 		DeletingPodSucceedsOrFails)
+	s.Step(`^pulling image "(.*)", logging in, and pushing local image to internal registry succeeds$`,
+		PullLoginTagPushImageSucceeds)
 
 	// CRC file operations
 	s.Step(`^file "([^"]*)" exists in CRC home folder$`,
@@ -662,4 +669,74 @@ func DeletingPodSucceedsOrFails(expected string) error {
 		err = util.ExecuteCommandSucceedsOrFails("oc delete pod $POD --now", expected)
 	}
 	return err
+}
+
+func PodmanCommandIsAvailable() error {
+
+	// Do what 'eval $(crc podman-env) would do
+	path := os.ExpandEnv("${HOME}/.crc/bin/oc:$PATH")
+	csshk := os.ExpandEnv("${HOME}/.crc/machines/crc/id_ecdsa")
+	dh := os.ExpandEnv("unix:///${HOME}/.crc/machines/crc/docker.sock")
+	ch := "ssh://core@127.0.0.1:2222/run/user/1000/podman/podman.sock"
+	if runtime.GOOS == "windows" {
+		userHomeDir, _ := os.UserHomeDir()
+		unexpandedPath := filepath.Join(userHomeDir, ".crc/bin/oc;${PATH}")
+		path = os.ExpandEnv(unexpandedPath)
+		csshk = filepath.Join(userHomeDir, ".crc/machines/crc/id_ecdsa")
+		dh = "npipe:////./pipe/rc-podman"
+	}
+	if runtime.GOOS == "linux" {
+		ch = "ssh://core@192.168.130.11:22/run/user/1000/podman/podman.sock"
+	}
+
+	os.Setenv("PATH", path)
+	os.Setenv("CONTAINER_SSHKEY", csshk)
+	os.Setenv("CONTAINER_HOST", ch)
+	os.Setenv("DOCKER_HOST", dh)
+
+	return nil
+
+}
+
+func ExecutingPodmanCommandSucceedsFails(command string, expected string) error {
+
+	var err error
+	if expected == "succeeds" {
+		_, err = cmd.RunPodmanExpectSuccess(strings.Split(command[1:len(command)-1], " ")...)
+	} else if expected == "fails" {
+		_, err = cmd.RunPodmanExpectFail(strings.Split(command[1:len(command)-1], " ")...)
+	}
+
+	return err
+}
+
+func PullLoginTagPushImageSucceeds(image string) error {
+	_, err := cmd.RunPodmanExpectSuccess("pull", image)
+	if err != nil {
+		return err
+	}
+
+	err = util.ExecuteCommand("oc whoami -t")
+	if err != nil {
+		return err
+	}
+
+	token := util.GetLastCommandOutput("stdout")
+	fmt.Println(token)
+	_, err = cmd.RunPodmanExpectSuccess("login", "-u", "kubeadmin", "-p", token, "default-route-openshift-image-registry.apps-crc.testing", "--tls-verify=false") // $(oc whoami -t)
+	if err != nil {
+		return err
+	}
+
+	_, err = cmd.RunPodmanExpectSuccess("tag", "quay.io/centos7/httpd-24-centos7", "default-route-openshift-image-registry.apps-crc.testing/testproj-img/hello:test")
+	if err != nil {
+		return err
+	}
+
+	_, err = cmd.RunPodmanExpectSuccess("push", "default-route-openshift-image-registry.apps-crc.testing/testproj-img/hello:test", "--tls-verify=false")
+	if err != nil {
+		return err
+	}
+
+	return nil
 }
